requirements.txt
Django==5.0
django-debug-toolbar==4.2.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environs==10.0.0
phonenumbers==8.13.27
Pillow==10.1.0
requests==2.31.0